requirements.txt
fastapi>=0.115.2,<1.0.0
uvicorn==0.34.0
streamlit>=1.47.0
python-multipart==0.0.7
python-docx==1.1.2
PyPDF2==3.0.1
scikit-learn==1.3.2
pandas==2.2.3
numpy==1.26.4
pydantic==2.10.6
requests>=2.32.3
boto3>=1.34.131
python-dotenv>=1.0.1
torch
transformers
sentence-transformers==2.2.2
langchain-community>=0.3.1
unstructured==0.10.30
packaging>=24.2
tokenizers<=0.20.3
langchain-core>=0.3.45,<1.0.0
huggingface-hub==0.25.2
docx2txt>=0.8
python-docx>=0.8.11
PyPDF2>=3.0.0
openai>=1.59.0
pdfplumber